requirements.txt
asgiref==3.7.2
Django==4.2.9
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
djangorestframework
djoser
django-rest-framework-nested
drf-nested-routers